requirements.txt
dune-client
pyyaml
python-dotenv
pandas
requests